requirements.txt
# Python >= 3.10 required
strands-agents>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0 
